Split SQL expression evaluation
